refactor: Improve the sign up screen
Move to using the same "Form => SingleChildScrollView => Column" approach I used for the Slam Book exercises to make text validation more robust and reliable. Also add validators for all the required fields in the sign up sheet.

While we're at it, improve messaging when Google account information is successfully obtained and log in automatically when a user already exists for a given account.

Signed-off-by: John Vincent Corcega <jmcorcega@up.edu.ph>