chore: Trigger sign out when sign up is cancelled
If Google Account is used to sign up, the user already exists in the provider, causing next restarts of the app to be completely broken as the sign up is unfinished. Ensure that we sign out when user cancels out of the sign up screen.

TODO: Trigger Firebase Auth to remove accounts as well when this happens.

Signed-off-by: John Vincent Corcega <jmcorcega@up.edu.ph>