refactor: Overhaul the user status and monitoring/quarantine dialogs
Overhaul the entire user details and 'confirm' dialogs to match the application's style and to streamline them into as fewer classes as possible to make it easier to style them. This also adds the functionality to add a user to the monitoring list when they report as 'sick'.

This commit also overhauls the entire user details modal to be easily glance-able and easier for an admin or building officer to add/remove a user to the monitoring/quarantine. Information tile have been made copy-able to the clipboard as well through a long tap.

Signed-off-by: John Vincent Corcega <jmcorcega@up.edu.ph>